requirements.txt
asgiref==3.10.0
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
razorpay==2.0.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.6.3
